Extract VarAnnotationProcessor from NodeScopeResolver
