+++ src/Analyser/StmtHandler/StaticVariableHandler.php
@@ -12,6 +12,7 @@ use PHPStan\Analyser\MutatingScope;
 use PHPStan\Analyser\NodeScopeResolver;
 use PHPStan\Analyser\StatementContext;
 use PHPStan\Analyser\StmtHandler;
+use PHPStan\Analyser\VarAnnotationProcessor;
 use PHPStan\DependencyInjection\AutowiredService;
 use PHPStan\ShouldNotHappenException;
 use PHPStan\TrinaryLogic;
@@ -25,6 +26,12 @@ use function is_string;
 #[AutowiredService]
 final class StaticVariableHandler implements StmtHandler
 {
+
+	public function __construct(
+		private VarAnnotationProcessor $varAnnotationProcessor,
+	)
+	{
+	}
 
 	public function supports(Stmt $stmt): bool
 	{
@@ -70,7 +77,7 @@ final class StaticVariableHandler implements StmtHandler
 			$vars[] = $var->var->name;
 		}
 
-		$scope = $nodeScopeResolver->processVarAnnotation($scope, $vars, $stmt);
+		$scope = $this->varAnnotationProcessor->processVarAnnotation($scope, $vars, $stmt);
 
 		return new InternalStatementResult($scope, hasYield: false, isAlwaysTerminating: false, exitPoints: [], throwPoints: [], impurePoints: $impurePoints);
 	}
